Thêm chức năng "Thêm sản phẩm"
