+++ BTLNhapMonCNPM/Areas/Admin/Controllers/ProductManagerController.cs
@@ -1,6 +1,7 @@
 ﻿using BTLNhapMonCNPM.Interface;
 using BTLNhapMonCNPM.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace BTLNhapMonCNPM.Areas.Admin.Controllers
 {
@@ -10,11 +11,13 @@ namespace BTLNhapMonCNPM.Areas.Admin.Controllers
     {
         private readonly LoaiSanPhamIT _danhmuc;
         private readonly SanPhamIT _sanpham;
-        public ProductManagerController(LoaiSanPhamIT danhmuc, SanPhamIT sanpham, NhaCCIT nhacc)
+		private readonly NhaCCIT _nhacc;
+		public ProductManagerController(LoaiSanPhamIT danhmuc, SanPhamIT sanpham, NhaCCIT nhacc)
         {
             _danhmuc = danhmuc;
             _sanpham = sanpham;
-        }
+			_nhacc = nhacc;
+		}
 
         
         
@@ -32,5 +35,26 @@ namespace BTLNhapMonCNPM.Areas.Admin.Controllers
             list["dsSP"] = listSP;
             return View(list);
         }
-    }
+
+
+		[HttpGet]
+		[Route("admin/product/add")]
+		public IActionResult Add()
+		{
+			var suppliers = _nhacc.getAllNhaCC();
+			var productTypes = _danhmuc.getAllDanhMuc();
+			ViewBag.Suppliers = new SelectList(suppliers, "IMaNcc", "STenNcc");
+			ViewBag.ProductTypes = new SelectList(productTypes, "IMaLoaiSp", "STenLoai");
+			return View(new TblSanPham());
+		}
+
+
+		[HttpPost]
+		[Route("admin/product/add")]
+		public IActionResult AddProduct(TblSanPham product)
+		{
+			_sanpham.AddSanPham(product);
+			return RedirectToAction("Index");
+		}
+	}
 }
