+++ BTLNhapMonCNPM/Areas/Admin/Views/ProductManager/Add.cshtml
@@ -1,0 +1,107 @@
+﻿@{
+    ViewData["Title"] = "Thêm sản phẩm";
+}
+
+<div class="container mt-4">
+    <div class="row justify-content-center">
+        <div class="col-md-8">
+            <div class="card">
+                <div class="card-header">
+                    <h3 class="mb-4">Thêm sản phẩm</h3>
+                </div>
+                <div class="card-body">
+                    <form asp-action="AddProduct" method="post" class="needs-validation" novalidate>
+                        <div class="mb-3">
+                            <label for="STen" class="form-label">Tên thuốc:</label>
+                            <input type="text" id="STen" name="STen" class="form-control" required />
+                            <div class="invalid-feedback">
+                                Hãy cung cấp tên thuốc
+                            </div>
+                        </div>
+                        <div class="mb-3">
+                            <label for="SDangThuoc" class="form-label">Dạng thuốc:</label>
+                            <input type="text" id="SDangThuoc" name="SDangThuoc" class="form-control" />
+                        </div>
+                        <div class="mb-3">
+                            <label for="SDonViTinh" class="form-label">Đơn vị tính:</label>
+                            <input type="text" id="SDonViTinh" name="SDonViTinh" class="form-control" />
+                        </div>
+                        <div class="mb-3">
+                            <label for="ISoLuong" class="form-label">Số lượng:</label>
+                            <input type="number" id="ISoLuong" name="ISoLuong" class="form-control" />
+                        </div>
+                        <div class="mb-3">
+                            <label for="SImageThuoc" class="form-label">URL hình ảnh:</label>
+                            <input type="text" id="SImageThuoc" name="SImageThuoc" class="form-control" />
+                        </div>
+                        <div class="mb-3">
+                            <label for="FGiaNhap" class="form-label">Giá nhập:</label>
+                            <input type="number" step="0.01" id="FGiaNhap" name="FGiaNhap" class="form-control" />
+                        </div>
+                        <div class="mb-3">
+                            <label for="FGiaBan" class="form-label">Giá bán:</label>
+                            <input type="number" step="0.01" id="FGiaBan" name="FGiaBan" class="form-control" />
+                        </div>
+                        <div class="mb-3">
+                            <label for="SCachSuDung" class="form-label">Cách sử dụng:</label>
+                            <input type="text" id="SCachSuDung" name="SCachSuDung" class="form-control" />
+                        </div>
+                        <div class="mb-3">
+                            <label for="SDieuKienBaoQuan" class="form-label">Điều kiện bảo quản:</label>
+                            <input type="text" id="SDieuKienBaoQuan" name="SDieuKienBaoQuan" class="form-control" />
+                        </div>
+                        <div class="mb-3">
+                            <label for="IMaNcc" class="form-label">Nhà cung cấp:</label>
+                            <select id="IMaNcc" name="IMaNcc" class="form-select" asp-items="ViewBag.Suppliers"></select>
+                        </div>
+                        <div class="mb-3">
+                            <label for="SHanSuDung" class="form-label">Hạn sử dụng:</label>
+                            <input type="text" id="SHanSuDung" name="SHanSuDung" class="form-control" />
+                        </div>
+                        <div class="mb-3 form-check">
+                            <input type="hidden" name="BBietTru" value="false" />
+                            <input type="checkbox" id="BBietTru" name="BBietTru" value="true" class="form-check-input" />
+                            <label for="BBietTru" class="form-check-label">Đặc biệt</label>
+                        </div>
+                        <div class="mb-3 form-check">
+                            <input type="hidden" name="BThuHoi" value="false" />
+                            <input type="checkbox" id="BThuHoi" name="BThuHoi" value="true" class="form-check-input" />
+                            <label for="BThuHoi" class="form-check-label">Thu hồi</label>
+                        </div>
+                        <div class="mb-3 form-check">
+                            <input type="hidden" name="BCanKeDon" value="false" />
+                            <input type="checkbox" id="BCanKeDon" name="BCanKeDon" value="true" class="form-check-input" />
+                            <label for="BCanKeDon" class="form-check-label">Cần kê đơn</label>
+                        </div>
+                        <div class="mb-3">
+                            <label for="IMaLoaiSp" class="form-label">Loại sản phẩm:</label>
+                            <select id="IMaLoaiSp" name="IMaLoaiSp" class="form-select" asp-items="ViewBag.ProductTypes"></select>
+                        </div>
+
+                        <button type="submit" class="btn btn-primary">Thêm sản phẩm</button>
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+
+<script>
+    (function () {
+        'use strict'
+
+        var forms = document.querySelectorAll('.needs-validation')
+
+        Array.prototype.slice.call(forms)
+            .forEach(function (form) {
+                form.addEventListener('submit', function (event) {
+                    if (!form.checkValidity()) {
+                        event.preventDefault()
+                        event.stopPropagation()
+                    }
+
+                    form.classList.add('was-validated')
+                }, false)
+            })
+    })()
+</script>
